🎨 Palette: Standardize slash-to-search shortcut across list pages
- Created reusable 'useSearchShortcut' hook for global '/' key listener.
- Refactored Experiments search to use the new hook.
- Added shortcut and visual '/' hint to Flags, Metrics, and Audit Log search inputs.
- Unified search input styling and accessibility (ARIA labels).
