+++ ui/src/app/metrics/page.tsx
@@ -1,7 +1,8 @@
 'use client';
 
-import { useEffect, useState, useCallback, useMemo } from 'react';
+import { useEffect, useState, useCallback, useMemo, useRef } from 'react';
 import dynamic from 'next/dynamic';
+import { useSearchShortcut } from '@/hooks/use-search-shortcut';
 import Link from 'next/link';
 import type { MetricDefinition, MetricType } from '@/lib/types';
 import { CompositeOperator } from '@/lib/types';
@@ -291,6 +292,8 @@ function MetricBrowserContent() {
   const [error, setError] = useState<string | null>(null);
   const [search, setSearch] = useState('');
   const [typeFilter, setTypeFilter] = useState<MetricType | ''>('');
+  const inputRef = useRef<HTMLInputElement>(null);
+  useSearchShortcut(inputRef);
 
   const hasActiveFilters = search !== '' || typeFilter !== '';
 
@@ -392,14 +395,20 @@ function MetricBrowserContent() {
             <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
           </svg>
           <input
+            ref={inputRef}
             type="text"
             placeholder="Search by name, ID, or description..."
             value={search}
             onChange={(e) => setSearch(e.target.value)}
-            className="w-full rounded-md border border-gray-300 py-1.5 pl-9 pr-3 text-sm shadow-sm focus:border-indigo-500 focus:outline-none focus:ring-1 focus:ring-indigo-500"
+            className="w-full rounded-md border border-gray-300 py-1.5 pl-9 pr-10 text-sm shadow-sm focus:border-indigo-500 focus:outline-none focus:ring-1 focus:ring-indigo-500"
             data-testid="metric-search"
             aria-label="Search metrics"
           />
+          <div className="pointer-events-none absolute right-3 top-1/2 flex -translate-y-1/2 items-center">
+            <span className="flex h-5 w-5 items-center justify-center rounded border border-gray-300 bg-gray-50 text-[10px] font-medium text-gray-500">
+              /
+            </span>
+          </div>
         </div>
         <select
           value={typeFilter}
